% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chi_test_cont_cpp}
\alias{chi_test_cont_cpp}
\title{run chi square test for continuous data.}
\usage{
chi_test_cont_cpp(dta, nbins = as.integer(c(100, 10)))
}
\arguments{
\item{dta}{A list of numeric vectors.}

\item{nbins}{A vector of length 2 of bin lengths.}
}
\value{
A list with test statistics and p values
}
\description{
run chi square test for continuous data.
}
