% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chi_test_disc_cpp}
\alias{chi_test_disc_cpp}
\title{run chi square test for discrete data.}
\usage{
chi_test_disc_cpp(dta, nbins = as.integer(c(100, 10)))
}
\arguments{
\item{dta}{A list of numeric vectors.}

\item{nbins}{Integer vector of length 2 with number of bins.}
}
\value{
A list with test statistics, p values and degrees of freedom
}
\description{
run chi square test for discrete data.
}
\keyword{internal}
