% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_cont_cpp}
\alias{permute_cont_cpp}
\title{permute continuous data}
\usage{
permute_cont_cpp(dta)
}
\arguments{
\item{dta}{A list of numeric vectors.}
}
\value{
A list of permuted x and y vectors.
}
\description{
permute continuous data
}
\keyword{internal}
