% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permute_disc_cpp}
\alias{permute_disc_cpp}
\title{permute discrete data}
\usage{
permute_disc_cpp(dta)
}
\arguments{
\item{dta}{A list of numeric vectors.}
}
\value{
A list of permuted vectors
}
\description{
permute discrete data
}
