% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_cpp}
\alias{power_cpp}
\title{Find the power of various tests via permutation.}
\usage{
power_cpp(
  rxy,
  nbins = as.integer(c(100, 10)),
  alpha = 0.05,
  B = 1000L,
  xparam = 0,
  yparam = 0,
  doMethod = as.character(c("chi large", "chi small", "t test", "KS", "Kuiper", "CvM",
    "AD", "LR", "ZA", "ZK", "ZC", "Wassp1"))
)
}
\arguments{
\item{rxy}{a function that generates x and y data.}

\item{nbins}{Two bin numbers for chi square  test.}

\item{alpha}{A numeric constant}

\item{B}{Number of simulation runs.}

\item{xparam}{arguments for r1.}

\item{yparam}{arguments for r2.}

\item{doMethod}{A character vector of methods to include}
}
\value{
A numeric matrix of powers
}
\description{
Find the power of various tests via permutation.
}
