% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rep_cpp}
\alias{rep_cpp}
\title{cpp version of R routine rep}
\usage{
rep_cpp(x, times)
}
\arguments{
\item{x}{numeric vector}

\item{times}{integer vector}
}
\value{
A numeric vector
}
\description{
cpp version of R routine rep
}
