% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_cont_cpp}
\alias{TS_cont_cpp}
\title{find test statistics for continuous data}
\usage{
TS_cont_cpp(
  dta,
  doMethod = as.character(c("chi large", "chi small", "t test", "KS", "Kuiper", "CvM",
    "AD", "LR", "ZA", "ZK", "ZC", "Wassp1"))
)
}
\arguments{
\item{dta}{A list}

\item{doMethod}{A character vector of methods to include}
}
\value{
A vector of test statistics
}
\description{
find test statistics for continuous data
}
