% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_disc_cpp}
\alias{TS_disc_cpp}
\title{find test statistics for discrete data}
\usage{
TS_disc_cpp(
  dta,
  ADweights,
  doMethod = as.character(c("chi large", "chi small", "t test", "KS", "Kuiper", "CvM",
    "AD", "LR", "ZA", "ZK", "ZC", "Wassp1"))
)
}
\arguments{
\item{dta}{A list}

\item{ADweights}{A vector of weights for AD method}

\item{doMethod}{A character vector of methods to include.}
}
\value{
A vector of test statistics
}
\description{
find test statistics for discrete data
}
\keyword{internal}
