% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weights_cpp}
\alias{weights_cpp}
\title{find weights for several statistics for discrete data}
\usage{
weights_cpp(dta)
}
\arguments{
\item{dta}{A list with vectors x, y and vals}
}
\value{
A vector of weights
}
\description{
find weights for several statistics for discrete data
}
